Prevent deprecation warning for PersistentDict: use PersistentMapping.
`persistent/dict.py`  simply has this code, for a long time now:

```
from zope.deferredimport import deprecated

deprecated(
    "`persistent.dict.PersistentDict` is deprecated. Use"
    " `persistent.mapping.PersistentMapping` instead."
    " This backward compatibility shim will be removed in persistent"
    " version 7.",
    PersistentDict='persistent.mapping:PersistentMapping',
)
```

Originally this was supposed to be removed in version 6.
And already since `persistent` 4.0.0 from 2012 is `PersistentDict` a backwards compatibility import for `PersistentMapping`.

So it seems safe to finally make the shift here. :-)